+++ lib/Component.lua
@@ -6,19 +6,35 @@
 
 local Reconciler = require(script.Parent.Reconciler)
 local Core = require(script.Parent.Core)
+local GlobalConfig = require(script.Parent.GlobalConfig)
+local Instrumentation = require(script.Parent.Instrumentation)
+
+local invalidSetStateMessages = require(script.Parent.invalidSetStateMessages)
 
 local Component = {}
 
+-- Locally cache tick so we can minimize impact of calling it for instrumentation
+local tick = tick
+
 Component.__index = Component
 
--- The error message that is thrown when setState is called in the wrong place.
--- This is declared here to avoid really messy indentation.
-local INVALID_SETSTATE_MESSAGE = [[
-setState cannot be used currently, are you calling setState from any of:
-* the willUpdate or willUnmount lifecycle hooks
-* the init function
-* the render function
-* the shouldUpdate function]]
+local function mergeState(currentState, partialState)
+	local newState = {}
+
+	for key, value in pairs(currentState) do
+		newState[key] = value
+	end
+
+	for key, value in pairs(partialState) do
+		if value == Core.None then
+			newState[key] = nil
+		else
+			newState[key] = value
+		end
+	end
+
+	return newState
+end
 
 --[[
 	Create a new Roact stateful component class.
@@ -52,9 +68,12 @@ function Component:extend(name)
 	function class._new(props, context)
 		local self = {}
 
+		-- When set to a value, setState will fail, using the given reason to
+		-- create a detailed error message.
+		-- You can see a list of reasons in invalidSetStateMessages.
+		self._setStateBlockedReason = nil
+
 		self.props = props
-		-- Used for tracking whether the component is in a position to set state.
-		self._canSetState = false
 		self._context = {}
 
 		-- Shallow copy all context values from our parent element.
@@ -68,7 +87,9 @@ function Component:extend(name)
 
 		-- Call the user-provided initializer, where state and _props are set.
 		if class.init then
+			self._setStateBlockedReason = "init"
 			class.init(self, props)
+			self._setStateBlockedReason = nil
 		end
 
 		-- The user constructer might not set state, so we can.
@@ -76,8 +97,13 @@ function Component:extend(name)
 			self.state = {}
 		end
 
-		-- Now that state has definitely been set, we can now allow it to be changed.
-		self._canSetState = true
+		if class.getDerivedStateFromProps then
+			local partialState = class.getDerivedStateFromProps(props, self.state)
+
+			if partialState then
+				self.state = mergeState(self.state, partialState)
+			end
+		end
 
 		return self
 	end
@@ -123,30 +149,32 @@ end
 	current state object.
 ]]
 function Component:setState(partialState)
-	-- State cannot be set in any lifecycle hooks.
-	if not self._canSetState then
-		error(INVALID_SETSTATE_MESSAGE, 0)
+	-- If setState was disabled, we should check for a detailed message and
+	-- display it.
+	if self._setStateBlockedReason ~= nil then
+		local messageSource = invalidSetStateMessages[self._setStateBlockedReason]
+
+		if messageSource == nil then
+			messageSource = invalidSetStateMessages["default"]
+		end
+
+		-- We assume that each message has a formatting placeholder for a component name.
+		local formattedMessage = string.format(messageSource, tostring(getmetatable(self)))
+
+		error(formattedMessage, 2)
 	end
 
 	-- If the partial state is a function, invoke it to get the actual partial state.
 	if type(partialState) == "function" then
 		partialState = partialState(self.state, self.props)
-	end
 
-	local newState = {}
-
-	for key, value in pairs(self.state) do
-		newState[key] = value
-	end
-
-	for key, value in pairs(partialState) do
-		if value == Core.None then
-			newState[key] = nil
-		else
-			newState[key] = value
+		-- If partialState is nil, abort the render.
+		if partialState == nil then
+			return
 		end
 	end
 
+	local newState = mergeState(self.state, partialState)
 	self:_update(self.props, newState)
 end
 
@@ -157,9 +185,22 @@ end
 	reconciliation step.
 ]]
 function Component:_update(newProps, newState)
-	self._canSetState = false
-	local doUpdate = self:shouldUpdate(newProps or self.props, newState or self.state)
-	self._canSetState = true
+	self._setStateBlockedReason = "shouldUpdate"
+
+	local doUpdate
+	if GlobalConfig.getValue("componentInstrumentation") then
+		-- Start timing
+		local time = tick()
+		doUpdate = self:shouldUpdate(newProps or self.props, newState or self.state)
+		-- Finish timing
+		time = tick() - time
+		-- Log result
+		Instrumentation.logShouldUpdate(self._handle, doUpdate, time)
+	else
+		doUpdate = self:shouldUpdate(newProps or self.props, newState or self.state)
+	end
+
+	self._setStateBlockedReason = nil
 
 	if doUpdate then
 		self:_forceUpdate(newProps, newState)
@@ -172,10 +213,26 @@ end
 	newProps and newState are optional.
 ]]
 function Component:_forceUpdate(newProps, newState)
-	self._canSetState = false
+	-- Compute new derived state.
+	-- Get the class - getDerivedStateFromProps is static.
+	local class = getmetatable(self)
+
+	-- Only update if newProps are given!
+	if newProps then
+		if class.getDerivedStateFromProps then
+			local derivedState = class.getDerivedStateFromProps(newProps, newState or self.state)
+
+			-- getDerivedStateFromProps can return nil if no changes are necessary.
+			if derivedState ~= nil then
+				newState = mergeState(newState or self.state, derivedState)
+			end
+		end
+	end
 
 	if self.willUpdate then
+		self._setStateBlockedReason = "willUpdate"
 		self:willUpdate(newProps or self.props, newState or self.state)
+		self._setStateBlockedReason = nil
 	end
 
 	local oldProps = self.props
@@ -189,8 +246,24 @@ function Component:_forceUpdate(newProps, newState)
 		self.state = newState
 	end
 
-	local newChildElement = self:render()
+	self._setStateBlockedReason = "render"
 
+	local newChildElement
+	if GlobalConfig.getValue("componentInstrumentation") then
+		-- Start timing
+		local time = tick()
+		newChildElement = self:render()
+		-- End timing
+		time = tick() - time
+		-- Log result
+		Instrumentation.logRenderTime(self._handle, time)
+	else
+		newChildElement = self:render()
+	end
+
+	self._setStateBlockedReason = nil
+
+	self._setStateBlockedReason = "reconcile"
 	if self._handle._reified ~= nil then
 		-- We returned an element before, update it.
 		self._handle._reified = Reconciler._reconcileInternal(
@@ -206,8 +279,7 @@ function Component:_forceUpdate(newProps, newState)
 			self._context
 		)
 	end
-
-	self._canSetState = true
+	self._setStateBlockedReason = nil
 
 	if self.didUpdate then
 		self:didUpdate(oldProps, oldState)
@@ -221,19 +293,57 @@ end
 function Component:_reify(handle)
 	self._handle = handle
 
-	local virtualTree = self:render()
+	self._setStateBlockedReason = "render"
+
+	local virtualTree
+	if GlobalConfig.getValue("componentInstrumentation") then
+		-- Start timing
+		local time = tick()
+		virtualTree = self:render()
+		-- End timing
+		time = tick() - time
+		-- Log result
+		Instrumentation.logRenderTime(self._handle, time)
+	else
+		virtualTree = self:render()
+	end
+
+	self._setStateBlockedReason = nil
+
 	if virtualTree then
+		self._setStateBlockedReason = "reconcile"
 		handle._reified = Reconciler._reifyInternal(
 			virtualTree,
 			handle._parent,
 			handle._key,
 			self._context
 		)
+		self._setStateBlockedReason = nil
 	end
 
 	if self.didMount then
 		self:didMount()
 	end
+end
+
+--[[
+	Destructs the component and invokes all necessary lifecycle methods.
+]]
+function Component:_teardown()
+	local handle = self._handle
+
+	if self.willUnmount then
+		self._setStateBlockedReason = "willUnmount"
+		self:willUnmount()
+		self._setStateBlockedReason = nil
+	end
+
+	-- Stateful components can return nil from render()
+	if handle._reified then
+		Reconciler.teardown(handle._reified)
+	end
+
+	self._handle = nil
 end
 
 return Component